requirements.txt
langchain>=0.1.0
langchain-community>=0.1.0
langgraph>=0.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
openai>=1.0.0
